rename release notes to releases where applicable
"Release Notes" is too long for the navigation bar. "Releases"
works well and is the same as the website. It also makes sense
to separate the releases screen from the actual release notes in
naming.
